rearrange to a separate module called crand.py
